Decoupled sound conditions (#1484)
* Convert sound conditions to a new format

* Convert animcommands to int

* Update flipeffect.cpp

* Update winmain.cpp

+++ TombEngine/Specific/level.cpp
@@ -388,7 +388,7 @@ void LoadObjects()
 
 	int commandCount = ReadInt32();
 	g_Level.Commands.resize(commandCount);
-	ReadBytes(g_Level.Commands.data(), sizeof(short) * commandCount);
+	ReadBytes(g_Level.Commands.data(), sizeof(int) * commandCount);
 
 	int boneCount = ReadInt32();
 	g_Level.Bones.resize(boneCount);
